Fix ethnews image asset source path and alt tag
GitHub pages is returning a 404 error for the ETHNews logo.

The casing of the URL didn't match the actual filename in the `/logos` directory. I also switched from the _.jpg_ to the _.png_ version already in that directory because it is ½ the size and has background transparency.

This commit corrects that and also updates the image alt attribute to reflect the same casing that ETHNews uses on their [About Us](https://www.ethnews.com/about-us) page.